adapt matrix SOn and add tests
